requirements.txt
fire
joblib
easydict
pandas
pytorch_lightning
scikit-learn
Cython
ray
tqdm
orjson
torch-model-archiver
